Fix race condition in System.Media.SoundPlayerTests
We were checking the IsLoadCompleted property immediately after our event handler run.

In some cases this could mean that the main test thread is signaled and running before
the worker thread has returned from posting the event, so _isLoadCompleted has not
yet been sent.

Workaround this by waiting for the internal semaphore to be signaled.  This is a bit
"too much information" for the test to couple to, but it proves the behavior and is
better than doing a thread.sleep.